third_party/WebKit/LayoutTests/http/tests/push_messaging/register_success.html

<!DOCTYPE html>
<html>
<head>
<title>register() succeeds when permission is granted and resolves with a valid registration</title>
<link rel="manifest" href="resources/push_manifest.json">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../serviceworker/resources/test-helpers.js"></script>
</head>
<body>
<script>
async_test(function(test) {
    var workerUrl = 'resources/empty_worker.js';
    var workerScope = 'resources/scope/' + location.pathname;
    var swRegistration;
    service_worker_unregister_and_register(test, workerUrl, workerScope)
        .then(function(serviceWorkerRegistration) {
            swRegistration = serviceWorkerRegistration;
            return wait_for_state(test, swRegistration.installing, 'activated');
        })
        .then(function() {
            // If running manually, grant permission when prompted.
            if (self.testRunner)
                testRunner.setPushMessagingPermission(location.origin, true);
            return swRegistration.pushManager.register();
        })
        .then(function(pushRegistration) {
            assert_own_property(pushRegistration, 'registrationId');
            assert_equals(typeof pushRegistration.registrationId, 'string');

            assert_own_property(pushRegistration, 'endpoint');
            assert_equals(typeof pushRegistration.endpoint, 'string');
            try {
              var endpointUrl = new URL(pushRegistration.endpoint);
            } catch(e) {
              assert_unreached('Constructing a URL from the endpoint should not throw.');
            }

            return service_worker_unregister_and_done(test, workerScope);
        })
        .catch(unreached_rejection(test));
}, 'register() succeeds when permission is granted and resolves with a valid registration');
</script>
</body>
</html>
